# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Fetchmark API
  version: "0.1.0"
  description: |
    Fetchmark is a self-hostable open-discovery API with explicit Tavily,
    Exa, and Brave compatibility routes. It brokers native open sources and
    SearXNG, fetches the returned URLs in parallel with
    SSRF-safe egress, extracts clean content with Trafilatura, and
    re-ranks results with BM25 + engine-diversity bonus.

    Authentication: every /v1/* endpoint requires either
    `Authorization: Bearer <api_key>` or `X-API-Key: <api_key>`.
    Fields flagged "admin only" require the caller's key to be in
    `FM_ADMIN_API_KEYS`; otherwise the request is rejected with 403.

    When the executable file resolved for the process at startup can be hashed,
    native `/v1/search` responses include `X-Fetchmark-Build-SHA256`. This is a
    build-artifact fingerprint, not a mapped-process-image, source-revision, or
    runtime-configuration digest.

    When startup configuration evidence is valid, native `/v1/search`
    responses also include `X-Fetchmark-Configuration-SHA256`. Authenticated
    clients can retrieve the exact bounded non-secret bytes from
    `/v1/evaluation/configuration`; the manifest explicitly identifies dynamic
    state that it cannot snapshot.
servers:
  - url: http://localhost:8080
    description: Local dev
security:
  - apiKeyHeader: []
  - apiKeyBearer: []
components:
  securitySchemes:
    apiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
    apiKeyBearer:
      type: http
      scheme: bearer
    exaApiKey:
      type: apiKey
      in: header
      name: x-api-key
    braveToken:
      type: apiKey
      in: header
      name: X-Subscription-Token
  headers:
    BuildSHA256:
      description: SHA-256 of the executable file resolved for the process at startup.
      required: false
      schema:
        type: string
        pattern: '^[0-9a-f]{64}$'
      example: '0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef'
    ConfigurationSHA256:
      description: >-
        SHA-256 of the exact startup-resolved non-secret evaluation
        configuration manifest returned by /v1/evaluation/configuration.
      required: false
      schema:
        type: string
        pattern: '^[0-9a-f]{64}$'
      example: 'abcdefabcdefabcdefabcdefabcdefabcdefabcdefabcdefabcdefabcdefabcd'
    DiscoveryAttribution:
      description: >-
        Optional RFC 8288 attribution link. Present when the returned search
        page includes a result supplied by an upstream whose terms request
        attribution, including Wiby, arXiv, and Stack Overflow. arXiv's Link
        relation is an acknowledgment; its CC0 identity stays scoped to native
        result metadata and does not license the mixed response.
      schema: { type: string }
      example: '<https://wiby.me/>; rel="via"; title="Wiby"'
  schemas:
...
    ExaRateLimitResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: { type: string, minLength: 1 }
paths:
...
  /v1/search:
    post:
      summary: Full pipeline — search + parallel fetch + extract + rank.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SearchRequest" }
      responses:
        "200":
          description: ok
          headers:
            X-Fetchmark-Build-SHA256: { $ref: "#/components/headers/BuildSHA256" }
            X-Fetchmark-Configuration-SHA256: { $ref: "#/components/headers/ConfigurationSHA256" }
            Link: { $ref: "#/components/headers/DiscoveryAttribution" }
          content:
            application/json:
              schema:
                type: object
                properties:
                  query:   { type: string }
                  count:   { type: integer }
                  results: { type: array, items: { $ref: "#/components/schemas/SearchResult" } }
                  discovery: { $ref: "#/components/schemas/DiscoveryReport" }
        "400":
          description: bad_request
          headers:
            X-Fetchmark-Build-SHA256: { $ref: "#/components/headers/BuildSHA256" }
            X-Fetchmark-Configuration-SHA256: { $ref: "#/components/headers/ConfigurationSHA256" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: missing_or_invalid_api_key
          headers:
            X-Fetchmark-Build-SHA256: { $ref: "#/components/headers/BuildSHA256" }
            X-Fetchmark-Configuration-SHA256: { $ref: "#/components/headers/ConfigurationSHA256" }
        "403":
          description: admin_only_flag_used_with_non_admin_key
          headers:
            X-Fetchmark-Build-SHA256: { $ref: "#/components/headers/BuildSHA256" }
            X-Fetchmark-Configuration-SHA256: { $ref: "#/components/headers/ConfigurationSHA256" }
        "429":
          description: rate_limited
          headers:
            X-Fetchmark-Build-SHA256: { $ref: "#/components/headers/BuildSHA256" }
            X-Fetchmark-Configuration-SHA256: { $ref: "#/components/headers/ConfigurationSHA256" }
        "502":
          description: search_failed; includes bounded discovery evidence when planning or provider execution began.
          headers:
            X-Fetchmark-Build-SHA256: { $ref: "#/components/headers/BuildSHA256" }
            X-Fetchmark-Configuration-SHA256: { $ref: "#/components/headers/ConfigurationSHA256" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "507":
          description: serialized response exceeds FM_MAX_REQUEST_OUTPUT_BYTES
          headers:
            X-Fetchmark-Build-SHA256: { $ref: "#/components/headers/BuildSHA256" }
            X-Fetchmark-Configuration-SHA256: { $ref: "#/components/headers/ConfigurationSHA256" }
  /v1/evaluation/configuration:
    get:
      summary: Return the exact startup-resolved non-secret evaluation configuration manifest.
      responses:
        "200":
          description: Exact manifest bytes whose digest is advertised on native search responses.
          headers:
            X-Fetchmark-Configuration-SHA256: { $ref: "#/components/headers/ConfigurationSHA256" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EvaluationConfigurationManifest" }
        "401": { description: missing_or_invalid_api_key }
        "429": { description: rate_limited }
        "503": { description: evaluation configuration unavailable because startup evidence is invalid }
  /v1/parse:
    post:
      summary: Parse one or more URLs in parallel (skips search).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ParseRequest" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:   { type: integer }
                  results: { type: array, items: { $ref: "#/components/schemas/SearchResult" } }
        "400": { description: urls_required }
        "401": { description: missing_or_invalid_api_key }
        "403": { description: admin_only_flag_used_with_non_admin_key }
        "429": { description: rate_limited }
        "507": { description: serialized response exceeds FM_MAX_REQUEST_OUTPUT_BYTES }
  /v1/summarize:
    post:
      summary: Summarize a single URL using an LLM provider (OpenAI- or Anthropic-compatible).
      description: |
        Parses the URL through the normal extraction pipeline, then asks
        the configured LLM to summarize the extracted text. The source
        page is wrapped in `<page>…</page>` delimiters and the model is
        instructed to treat it as untrusted. Raw reasoning/thinking text
        is never returned — only token counts.

        Non-admin callers may only override provider, model, or thinking
        when the corresponding server config gate allows it. max_tokens,
        timeout_ms, instructions length, and thinking.budget_tokens caps
        apply to all callers, including admins.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:          { type: string, format: uri }
                format:       { type: string, enum: [markdown, plain, bullets], default: markdown }
                instructions: { type: string }
                provider:     { type: string, description: "Provider name (not kind). Defaults to the configured default." }
                model:        { type: string }
                max_tokens:   { type: integer }
                temperature:  { type: number }
                timeout_ms:   { type: integer }
                render:       { type: boolean }
                thinking:
                  type: object
                  properties:
                    enabled:       { type: boolean }
                    budget_tokens: { type: integer, description: "Anthropic only, >= 1024 when set" }
                    effort:        { type: string, enum: [low, medium, high], description: "OpenAI reasoning_effort" }
      responses:
        "200":
          description: summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:  { type: string }
                  provider: { type: string }
                  model:    { type: string }
                  url:      { type: string }
                  title:    { type: string }
                  usage:
                    type: object
                    properties:
                      prompt_tokens:     { type: integer }
                      completion_tokens: { type: integer }
                      reasoning_tokens:  { type: integer }
                      total_tokens:      { type: integer }
                  source:
                    type: object
                    properties:
                      title:          { type: string }
                      author:         { type: string }
                      site_name:      { type: string }
                      from_cache:     { type: boolean }
                      word_count:     { type: integer, description: "Word count of the included source body sent for summarization." }
                      truncated:      { type: boolean, description: "True when Fetchmark truncated extracted content before sending it to the provider." }
                      original_chars: { type: integer, description: "Rune count of the extracted source body before truncation." }
                      included_chars: { type: integer, description: "Rune count included in the summarization prompt." }
        "400": { description: bad_request }
        "401": { description: missing_or_invalid_api_key }
        "422": { description: unsupported_or_empty_content }
        "429": { description: provider_rate_limited }
        "502": { description: provider_auth_or_upstream_error }
        "503": { description: summarize_not_configured }
        "504": { description: provider_timeout }
  /compat/tavily/search:
    post:
      summary: Tavily-compatible open-search subset.
      security: [{ apiKeyBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TavilySearchRequest" }
      responses:
        "200":
          description: Tavily-shaped search response
          headers:
            Link: { $ref: "#/components/headers/DiscoveryAttribution" }
        "400": { description: Unsupported or invalid Tavily request }
        "401": { description: Missing or invalid Bearer key }
        "429": { description: Tavily-shaped rate limit response }
        "507": { description: Tavily-shaped response budget error }
  /compat/exa/search:
    post:
      summary: Exa-compatible open-search subset.
      security: [{ exaApiKey: [] }, { apiKeyBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExaSearchRequest" }
      responses:
        "200":
          description: Exa-shaped search response
          headers:
            Link: { $ref: "#/components/headers/DiscoveryAttribution" }
        "400":
          description: Invalid body, conflicting controls, invalid numResults, or instance result cap exceeded.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExaErrorResponse" }
        "401":
          description: Missing, empty, or invalid key (`INVALID_API_KEY`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExaErrorResponse" }
        "429":
          description: Exa-compatible simple rate-limit response; contains only `error`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExaRateLimitResponse" }
        "507":
          description: Fetchmark-only response-byte-budget extension (`RESPONSE_TOO_LARGE`), not an Exa status/tag.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExaErrorResponse" }
  /compat/exa/contents:
    post:
      summary: Exa-compatible URL content retrieval.
      security: [{ exaApiKey: [] }, { apiKeyBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExaContentsRequest" }
      responses:
        "200": { description: Results plus per-URL statuses }
        "400":
          description: Invalid body, conflicting controls, or invalid URLs/IDs.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExaErrorResponse" }
        "401":
          description: Missing, empty, or invalid key (`INVALID_API_KEY`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExaErrorResponse" }
        "429":
          description: Exa-compatible simple rate-limit response; contains only `error`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExaRateLimitResponse" }
        "507":
          description: Fetchmark-only response-byte-budget extension (`RESPONSE_TOO_LARGE`), not an Exa status/tag.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExaErrorResponse" }
  /compat/brave/res/v1/web/search:
    get:
      summary: Brave-compatible web-search subset.
      security: [{ braveToken: [] }]
      parameters:
        - { in: query, name: q, required: true, schema: { type: string, maxLength: 400 } }
        - { in: query, name: count, schema: { type: integer, minimum: 1, maximum: 20, default: 20 } }
        - { in: query, name: offset, schema: { type: integer, minimum: 0, maximum: 9, default: 0 } }
        - { in: query, name: search_lang, schema: { type: string, default: en } }
        - { in: query, name: safesearch, schema: { type: string, enum: [off, moderate, strict], default: moderate } }
        - { in: query, name: freshness, schema: { type: string, enum: [pd, pm, py] } }
        - { in: query, name: spellcheck, schema: { type: boolean } }
      responses:
        "200":
          description: Brave-shaped web results
          headers:
            Link: { $ref: "#/components/headers/DiscoveryAttribution" }
        "422": { description: Unsupported or invalid Brave query }
        "401": { description: Missing or invalid subscription token }
        "429": { description: Brave-shaped rate limit response }
        "507": { description: Brave-shaped response budget error }
  /admin/summarize/config:
    get:
      summary: (admin) Snapshot of configured providers. API keys are never returned.
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  default:   { type: string }
                  providers: { type: array, items: { type: object } }
        "401": { description: missing_or_invalid_admin_key }
  /admin/corpus/admissions:
    post:
      summary: (admin) Fresh-fetch and explicitly admit URL-only batches into a curated local corpus.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [urls]
              properties:
                urls:
                  type: array
                  minItems: 1
                  items: { type: string, format: uri, maxLength: 2048 }
                safety_classification:
                  type: string
                  enum: [unclassified, safe, unsafe]
                  default: unclassified
                  description: Explicit operator assertion applied to every admitted URL; Fetchmark does not infer this from page text.
      responses:
        "200": { description: "Bodyless per-URL admitted, rejected, or failed outcomes" }
        "400": { description: "Invalid, duplicate, or oversized URL batch" }
        "401": { description: Missing or invalid admin key }
        "409": { description: Curated mode is not active }
        "503": { description: Curated stores are not ready }
        "507": { description: Response byte budget exceeded }
  /admin/corpus/focused-admissions:
    post:
      summary: (admin) Admit one focused-crawler URL under an explicit redirect path policy.
      description: >-
        Curated-only control plane. Fetchmark fresh-fetches the URL and requires
        every redirect to retain the original scheme and authority, match an
        allowed decoded-path prefix, and miss all denied paths and canonical
        URLs. The request carries no page body and cannot assert safety. A
        bounded crawler-only outbound-link snapshot may be requested; it is
        returned only after successful permitted retention.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [url, allowed_path_prefixes]
              properties:
                url: { type: string, format: uri, maxLength: 2048 }
                allowed_path_prefixes:
                  type: array
                  minItems: 1
                  maxItems: 256
                  uniqueItems: true
                  items: { type: string, maxLength: 2048, pattern: '^/' }
                denied_path_prefixes:
                  type: array
                  maxItems: 256
                  uniqueItems: true
                  items: { type: string, maxLength: 2048, pattern: '^/' }
                denied_urls:
                  type: array
                  maxItems: 256
                  uniqueItems: true
                  items: { type: string, format: uri, maxLength: 2048 }
                max_outbound_links:
                  type: integer
                  minimum: 0
                  maximum: 64
                  default: 0
                  description: >-
                    Crawler-only reader-view link cap. Zero omits link metadata;
                    a requested admitted response contains outbound_links even
                    when the authoritative snapshot is empty.
      responses:
        "200":
          description: One admitted, rejected, or failed outcome without page content.
          content:
            application/json:
              schema:
                type: object
                required: [count, results]
                properties:
                  count: { type: integer, enum: [1] }
                  results:
                    type: array
                    minItems: 1
                    maxItems: 1
                    items:
                      type: object
                      additionalProperties: false
                      required: [url, status]
                      properties:
                        url: { type: string, format: uri, maxLength: 2048 }
                        status: { type: string, enum: [admitted, rejected, failed] }
                        reason: { type: string }
                        outbound_links:
                          type: array
                          maxItems: 64
                          uniqueItems: true
                          items: { type: string, format: uri, maxLength: 2048 }
        "400": { description: "Invalid URL, ambiguous path, or invalid/oversized policy" }
        "401": { description: Missing or invalid admin key }
        "409": { description: Curated mode is not active }
        "503": { description: Curated stores are not ready }
        "507": { description: Response byte budget exceeded }
  /admin/corpus/takedowns:
    post:
      summary: (admin) Write sticky curated/archive tombstones and purge all retained source versions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [urls]
              properties:
                urls:
                  type: array
                  minItems: 1
                  items: { type: string, format: uri, maxLength: 2048 }
      responses:
        "200": { description: Bodyless per-URL taken_down or failed outcomes }
        "400": { description: "Invalid, duplicate, or oversized URL batch" }
        "401": { description: Missing or invalid admin key }
        "409": { description: Curated or archive mode is not active }
        "503": { description: Persistent corpus stores are not ready }
        "507": { description: Response byte budget exceeded }
  /admin/summarize/providers:
    put:
      summary: (admin) Upsert a single provider. Omitted fields preserve prior values.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:        { type: string }
                kind:        { type: string, enum: [openai, anthropic] }
                base_url:    { type: string }
                api_key:     { type: string, description: "Write-only. Never returned." }
                model:       { type: string }
                timeout_ms:  { type: integer }
                max_tokens:  { type: integer }
                temperature: { type: number }
                thinking:    { type: object }
      responses:
        "200": { description: updated }
        "400": { description: invalid_config }
        "401": { description: missing_or_invalid_admin_key }
  /admin/summarize/providers/{name}:
    delete:
      summary: (admin) Remove a provider by name.
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
      responses:
        "200": { description: deleted }
        "401": { description: missing_or_invalid_admin_key }
  /admin/summarize/default:
    put:
      summary: (admin) Set the default provider by name.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        "200": { description: ok }
        "400": { description: unknown_provider }
        "401": { description: missing_or_invalid_admin_key }